+++ packages/ifpack2/src/Ifpack2_Details_Chebyshev_decl.hpp
@@ -656,46 +656,6 @@ private:
                    const ST eigRatio,
                    const V& D_inv);
 
-  /// \brief Fill x with random initial guess for power method
-  ///
-  /// \param x [out] Initial guess vector; a domain Map vector of the
-  ///   matrix.
-  /// \param nonnegativeRealParts [in] Whether to force all entries of
-  ///   x (on output) to have nonnegative real parts.  Defaults to
-  ///   false (don't force).
-  ///
-  /// This is an implementation detail of powerMethod() below.  For a
-  /// justification of the second parameter, see Github Issues #64 and
-  /// #567.
-  void computeInitialGuessForPowerMethod (V& x, const bool nonnegativeRealParts = false) const;
-
-  /// \brief Use the power method to estimate the maximum eigenvalue
-  ///   of A*D_inv, given an initial guess vector x.
-  ///
-  /// \param A [in] The Operator to use.
-  /// \param D_inv [in] Vector to use as implicit right scaling of A.
-  /// \param numIters [in] Maximum number of iterations of the power
-  ///   method.
-  /// \param x [in/out] On input: Initial guess Vector for the power
-  ///   method.  Its Map must be the same as that of the domain Map of
-  ///   A.  This method may use this Vector as scratch space.
-  ///
-  /// \return Estimate of the maximum eigenvalue of A*D_inv.
-  ST
-  powerMethodWithInitGuess (const op_type& A, const V& D_inv, const int numIters, V& x);
-
-  /// \brief Use the power method to estimate the maximum eigenvalue
-  ///   of A*D_inv.
-  ///
-  /// \param A [in] The Operator to use.
-  /// \param D_inv [in] Vector to use as implicit right scaling of A.
-  /// \param numIters [in] Maximum number of iterations of the power
-  ///   method.
-  ///
-  /// \return Estimate of the maximum eigenvalue of A*D_inv.
-  ST
-  powerMethod (const op_type& A, const V& D_inv, const int numIters);
-
   /// \brief Use the cg method to estimate the maximum eigenvalue
   ///   of A*D_inv, given an initial guess vector x.
   ///
